requirements.txt
tensorflow==2.1.0
tqdm>=4.26.1
PyYAML>=3.12
